requirements.txt
fastapi==0.116.1
jinja2==3.1.6
numpy==2.3.4
pandas==2.3.3
python-dotenv==1.1.1
requests==2.32.5
uvicorn[standard]==0.35.0
